refactor smtp 163 or 126 send mail by mail.channel=local
