--- src/main/resources/email.properties
@@ -1,28 +1,0 @@
-#config for SMTP email 
-#author: snowflake
-#version 1.0.0.0
-#since 2.0.0
-#\u914d\u7f6e\u542f\u7528smtp\u53d1\u90ae\u4ef6\u7684\u529f\u80fd\uff0c\u5c06\u4f1a\u8986\u76d6Mails\u7684\u8bbe\u7f6e\uff0c\u800c\u4f7f\u7528MailSender\u7684\u6a21\u5f0f
-smtp.enable=true
-#\u8bbe\u7f6e\u4e3adebug\u6a21\u5f0f\u65f6\uff0c\u540e\u53f0\u63a7\u5236\u53f0\u4f1a\u8f93\u51fa\u90ae\u4ef6\u53d1\u9001\u8fc7\u7a0b
-isdebug=true
-#\u4f20\u8f93\u534f\u8bae\u4f7f\u7528smtp
-mail.transport.protocol=smtp
-#smtp\u90ae\u4ef6\u670d\u52a1\u5668
-mail.host=smtp.163.com
-#smtp\u7aef\u53e3\u53f7\u8bbe\u7f6e
-mail.port=25
-
-#\u5f00\u542f\u8ba4\u8bc1
-mail.smtp.auth=true
-mail.smtp.ssl=true
-mail.smtp.starttls.enable=true
-
-#\u53d1\u9001\u8005\u7684Email
-mail.smtp.sender=snowflake3721@163.com
-#\u53d1\u9001\u8005\u540d\u79f0
-mail.smtp.username = snowflake3721
-#\u90ae\u4ef6\u9a8c\u8bc1\u5bc6\u7801\uff0c126\u7684STMP\u670d\u52a1\u7684\u5ba2\u6237\u7aef\u5bc6\u7801\u9700\u8981\u81ea\u5df1\u5f00\u901a\uff0c\u4e0d\u662f\u767b\u5f55\u5bc6\u7801\u54e6\uff0c\u6362\u6210\u4f60\u81ea\u5df1\u7684
-mail.smtp.passsword=snowflake3721
-#\u90ae\u4ef6\u4fdd\u5b58\u76ee\u5f55\uff0c\u7edd\u5bf9\u8def\u5f84
-saved.eml.path= D:/
